feat: migrate frontend ESLint to flat config and create shared base
- Create eslint-base.config.mjs with shared rules and ignore patterns
- Update backend/eslint.config.mjs to import from shared base
- Update frontend/package.json: remove legacy configs (eslint-config-love, import resolvers, promise, import plugins)
- Add ESLint 9 flat config dependencies (@eslint/eslintrc, @eslint/js, globals)
- Update ESLint and TypeScript ESLint to match root versions (9.16.0, 8.16.0)
- Create frontend/eslint.config.mjs with React support and import from base
- Preserve all existing React rules from .eslintrc.yml
- Remove frontend/.eslintrc.yml and frontend/.eslintignore
- Split lint scripts: lint (check-only) and lint:fix (auto-fix)
- Update package-lock.json files

Part of multi-PR ESLint migration plan.
