--- frontend/.eslintignore
@@ -1,7 +1,0 @@
-.husky/
-.vscode/
-.yarn/
-coverage/
-dist/
-public/assets/
-tsconfig.*.json
