+++ frontend/package.json
@@ -12,7 +12,9 @@
     "preview": "vite preview",
     "test:unit": "vitest --mode test",
     "test:cov": "vitest run --mode test --coverage",
-    "build": "vite build"
+    "build": "vite build",
+    "lint": "eslint \"src/**/*.{ts,tsx}\" \"e2e/**/*.ts\"",
+    "lint:fix": "eslint \"src/**/*.{ts,tsx}\" \"e2e/**/*.ts\" --fix"
   },
   "dependencies": {
     "@bcgov/bc-sans": "^2.1.0",
@@ -43,19 +45,19 @@
     "@types/react-dom": "^19.0.4",
     "@vitest/coverage-v8": "^3.0.8",
     "@vitest/ui": "^3.0.8",
-    "eslint": "^9.2.2",
-    "eslint-config-love": "^133.0.0",
-    "eslint-config-prettier": "^10.1.1",
-    "eslint-import-resolver-alias": "^1.1.2",
-    "eslint-import-resolver-typescript": "^4.0.0",
-    "eslint-plugin-cypress": "^5.0.0",
+    "@eslint/eslintrc": "^3.2.0",
+    "@eslint/js": "^9.15.0",
+    "@typescript-eslint/eslint-plugin": "^8.16.0",
+    "@typescript-eslint/parser": "^8.16.0",
+    "eslint": "^9.16.0",
+    "eslint-config-prettier": "^10.0.0",
+    "eslint-plugin-cypress": "^4.2.0",
     "eslint-plugin-html": "^8.1.2",
-    "eslint-plugin-import": "^2.31.0",
-    "eslint-plugin-prettier": "^5.2.3",
-    "eslint-plugin-promise": "^7.2.1",
+    "eslint-plugin-prettier": "^5.2.1",
     "eslint-plugin-react": "^7.37.4",
     "eslint-plugin-react-hooks": "^7.0.0",
     "eslint-plugin-tsdoc": "^0.4.0",
+    "globals": "^16.0.0",
     "history": "^5.3.0",
     "jsdom": "^27.0.0",
     "msw": "^2.7.3",
